app/requirements.txt
asgiref==3.6.0
daphne
twisted[http2,tls]
gunicorn
websocket-client
django-redis
django-ratelimit
python-slugify
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
daphne==4.0.0
Django==4.1.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
oauthlib==3.2.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.15
zope.interface==6.0
dj-database-url
